+++ swap/src/lib.rs
@@ -27,5 +27,6 @@ pub mod monero;
 pub mod network;
 pub mod protocol;
 pub mod seed;
+pub mod tor;
 
 mod monero_ext;